requirements.txt
flask
llama-index
sentence-transformers
pinecone-client
requests